feat: implement ActionCableConnectionSocket service for WebSocket handling
