+++ package.json
@@ -62,7 +62,6 @@
     "@zxing/browser": "0.1.5",
     "@zxing/library": "0.21.3",
     "@zxing/ngx-scanner": "19.0.0",
-
     "angular2-multiselect-dropdown": "^5.0.4",
     "angular2-smart-table": "^3.1.1",
     "autolinker": "^4.0.0",
@@ -129,7 +128,6 @@
     "@tailwindcss/aspect-ratio": "^0.4.2",
     "@tailwindcss/typography": "^0.5.10",
     "@twittwer/compodoc": "^1.13.0",
-
     "@types/archiver": "~6.0.2",
     "@types/canvas-confetti": "~1.6.4",
     "@types/chart.js": "~2.9.41",
